Add overwrite flag to media downloads

+++ instagrapi/mixins/photo.py
@@ -43,7 +43,9 @@ class DownloadPhotoMixin:
     Helpers for downloading photo
     """
 
-    def photo_download(self, media_pk: int, folder: Path = "") -> Path:
+    def photo_download(
+        self, media_pk: int, folder: Path = "", overwrite: bool = True
+    ) -> Path:
         """
         Download photo using media pk
 
@@ -54,6 +56,9 @@ class DownloadPhotoMixin:
         folder: Path, optional
             Directory in which you want to download the photo, default is "" and will download the files to working
                 directory
+        overwrite: bool, optional
+            Whether to overwrite an existing file. When False and the target path already exists, skip download and
+                return the existing path.
 
         Returns
         -------
@@ -65,10 +70,16 @@ class DownloadPhotoMixin:
         filename = "{username}_{media_pk}".format(
             username=media.user.username, media_pk=media_pk
         )
-        return self.photo_download_by_url(media.thumbnail_url, filename, folder)
+        return self.photo_download_by_url(
+            media.thumbnail_url, filename, folder, overwrite=overwrite
+        )
 
     def photo_download_by_url(
-        self, url: str, filename: str = "", folder: Path = ""
+        self,
+        url: str,
+        filename: str = "",
+        folder: Path = "",
+        overwrite: bool = True,
     ) -> Path:
         """
         Download photo using URL
@@ -82,6 +93,9 @@ class DownloadPhotoMixin:
         folder: Path, optional
             Directory in which you want to download the photo, default is "" and will download the files to working
                 directory
+        overwrite: bool, optional
+            Whether to overwrite an existing file. When False and the target path already exists, skip download and
+                return the existing path.
 
         Returns
         -------
@@ -92,6 +106,8 @@ class DownloadPhotoMixin:
         fname = urlparse(url).path.rsplit("/", 1)[1]
         filename = "%s.%s" % (filename, fname.rsplit(".", 1)[1]) if filename else fname
         path = Path(folder) / filename
+        if path.exists() and not overwrite:
+            return path.resolve()
         response = requests.get(url, stream=True, timeout=self.request_timeout)
         response.raise_for_status()
         with open(path, "wb") as f:
